Separate ifft
