+++ .vscode/settings.json
@@ -20,5 +20,8 @@
                 }
             }
         ]
-    }
+    },
+    "cSpell.words": [
+        "itwiddle"
+    ]
 }
