+++ src/core/backend/avx512/circle.rs
@@ -1,5 +1,5 @@
 use super::{as_cpu_vec, AVX512Backend};
-use crate::core::backend::{CPUBackend, Column};
+use crate::core::backend::{CPUBackend, Col};
 use crate::core::fields::m31::BaseField;
 use crate::core::poly::circle::{
     CanonicCoset, CircleDomain, CircleEvaluation, CirclePoly, PolyOps,
@@ -8,19 +8,16 @@ use crate::core::poly::circle::{
 impl PolyOps<BaseField> for AVX512Backend {
     fn new_canonical_ordered(
         coset: CanonicCoset,
-        values: Column<Self, BaseField>,
+        values: Col<Self, BaseField>,
     ) -> CircleEvaluation<Self, BaseField> {
         let eval = CPUBackend::new_canonical_ordered(coset, as_cpu_vec(values));
-        CircleEvaluation::new(
-            eval.domain,
-            Column::<AVX512Backend, _>::from_iter(eval.values),
-        )
+        CircleEvaluation::new(eval.domain, Col::<AVX512Backend, _>::from_iter(eval.values))
     }
 
     fn interpolate(eval: CircleEvaluation<Self, BaseField>) -> CirclePoly<Self, BaseField> {
         let cpu_eval = CircleEvaluation::<CPUBackend, _>::new(eval.domain, as_cpu_vec(eval.values));
         let cpu_poly = cpu_eval.interpolate();
-        CirclePoly::new(Column::<AVX512Backend, _>::from_iter(cpu_poly.coeffs))
+        CirclePoly::new(Col::<AVX512Backend, _>::from_iter(cpu_poly.coeffs))
     }
 
     fn eval_at_point<E: crate::core::fields::ExtensionOf<BaseField>>(
@@ -40,13 +37,13 @@ impl PolyOps<BaseField> for AVX512Backend {
         let cpu_eval = cpu_poly.evaluate(domain);
         CircleEvaluation::new(
             cpu_eval.domain,
-            Column::<AVX512Backend, _>::from_iter(cpu_eval.values),
+            Col::<AVX512Backend, _>::from_iter(cpu_eval.values),
         )
     }
 
     fn extend(poly: &CirclePoly<Self, BaseField>, log_size: u32) -> CirclePoly<Self, BaseField> {
         let cpu_poly = CirclePoly::<CPUBackend, _>::new(as_cpu_vec(poly.coeffs.clone()));
-        CirclePoly::new(Column::<AVX512Backend, _>::from_iter(
+        CirclePoly::new(Col::<AVX512Backend, _>::from_iter(
             cpu_poly.extend(log_size).coeffs,
         ))
     }
